[codex-login] isolate HTTP state tests [ci changed_files]
