+++ codex-rs/login/src/http_state.rs
@@ -28,42 +28,5 @@ pub(crate) fn clear_all(codex_home: &Path) {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use tempfile::TempDir;
-
-    #[test]
-    fn replace_after_login_sets_and_clears_surface_state() {
-        let codex_home = TempDir::new().expect("tempdir");
-        let store = HttpStateStore::new(codex_home.path().to_path_buf());
-        store
-            .set(HttpStateSurface::CodexCli, "stale-cli-state".to_string())
-            .expect("CLI state should store");
-
-        replace_after_login(
-            codex_home.path(),
-            HttpStateSurface::CodexDesktop,
-            Some("minted-state".to_string()),
-        );
-        assert_eq!(
-            store
-                .get(HttpStateSurface::CodexCli)
-                .expect("CLI state should load"),
-            None,
-        );
-        assert_eq!(
-            store
-                .get(HttpStateSurface::CodexDesktop)
-                .expect("state should load"),
-            Some("minted-state".to_string()),
-        );
-
-        replace_after_login(codex_home.path(), HttpStateSurface::CodexDesktop, None);
-        assert_eq!(
-            store
-                .get(HttpStateSurface::CodexDesktop)
-                .expect("state should load"),
-            None,
-        );
-    }
-}
+#[path = "http_state_tests.rs"]
+mod tests;
