+++ codex-rs/login/tests/suite/login_server_e2e.rs
@@ -26,7 +26,10 @@ const WORKSPACE_ID_DISALLOWED: &str = "123e4567-e89b-42d3-a456-426614174002";
 
 // See spawn.rs for details
 
-fn start_mock_issuer(chatgpt_account_id: &str) -> (SocketAddr, thread::JoinHandle<()>) {
+fn start_mock_issuer(
+    chatgpt_account_id: &str,
+    http_state_surface: HttpStateSurface,
+) -> (SocketAddr, thread::JoinHandle<()>) {
     // Bind to a random available port
     let listener = TcpListener::bind(("127.0.0.1", 0)).unwrap();
     let addr = listener.local_addr().unwrap();
@@ -40,6 +43,12 @@ fn start_mock_issuer(chatgpt_account_id: &str) -> (SocketAddr, thread::JoinHandl
                 // Read body
                 let mut body = String::new();
                 let _ = req.as_reader().read_to_string(&mut body);
+                if body.contains("grant_type=authorization_code") {
+                    assert!(
+                        body.contains(&format!("surface={http_state_surface}")),
+                        "authorization-code exchange should identify the initiating surface"
+                    );
+                }
                 // Build minimal JWT with plan=pro
                 #[derive(serde::Serialize)]
                 struct Header {
@@ -95,7 +104,8 @@ async fn end_to_end_login_flow_persists_auth_json() -> Result<()> {
     skip_if_no_network!(Ok(()));
 
     let chatgpt_account_id = "12345678-0000-0000-0000-000000000000";
-    let (issuer_addr, issuer_handle) = start_mock_issuer(chatgpt_account_id);
+    let (issuer_addr, issuer_handle) =
+        start_mock_issuer(chatgpt_account_id, HttpStateSurface::CodexDesktop);
     let issuer = format!("http://{}:{}", issuer_addr.ip(), issuer_addr.port());
 
     let tmp = tempdir()?;
@@ -178,7 +188,8 @@ async fn end_to_end_login_flow_persists_auth_json() -> Result<()> {
 async fn creates_missing_codex_home_dir() -> Result<()> {
     skip_if_no_network!(Ok(()));
 
-    let (issuer_addr, _issuer_handle) = start_mock_issuer(WORKSPACE_ID_ALLOWED);
+    let (issuer_addr, _issuer_handle) =
+        start_mock_issuer(WORKSPACE_ID_ALLOWED, HttpStateSurface::CodexCli);
     let issuer = format!("http://{}:{}", issuer_addr.ip(), issuer_addr.port());
 
     let tmp = tempdir()?;
@@ -222,7 +233,8 @@ async fn creates_missing_codex_home_dir() -> Result<()> {
 async fn login_server_includes_forced_workspaces_as_one_query_param() -> Result<()> {
     skip_if_no_network!(Ok(()));
 
-    let (issuer_addr, _issuer_handle) = start_mock_issuer(WORKSPACE_ID_ALLOWED);
+    let (issuer_addr, _issuer_handle) =
+        start_mock_issuer(WORKSPACE_ID_ALLOWED, HttpStateSurface::CodexCli);
     let issuer = format!("http://{}:{}", issuer_addr.ip(), issuer_addr.port());
 
     let tmp = tempdir()?;
@@ -264,7 +276,8 @@ async fn login_server_includes_forced_workspaces_as_one_query_param() -> Result<
 async fn forced_chatgpt_workspace_id_mismatch_blocks_login() -> Result<()> {
     skip_if_no_network!(Ok(()));
 
-    let (issuer_addr, _issuer_handle) = start_mock_issuer(WORKSPACE_ID_DISALLOWED);
+    let (issuer_addr, _issuer_handle) =
+        start_mock_issuer(WORKSPACE_ID_DISALLOWED, HttpStateSurface::CodexCli);
     let issuer = format!("http://{}:{}", issuer_addr.ip(), issuer_addr.port());
 
     let tmp = tempdir()?;
@@ -325,7 +338,8 @@ async fn forced_chatgpt_workspace_id_mismatch_blocks_login() -> Result<()> {
 async fn oauth_access_denied_missing_entitlement_blocks_login_with_clear_error() -> Result<()> {
     skip_if_no_network!(Ok(()));
 
-    let (issuer_addr, _issuer_handle) = start_mock_issuer(WORKSPACE_ID_ALLOWED);
+    let (issuer_addr, _issuer_handle) =
+        start_mock_issuer(WORKSPACE_ID_ALLOWED, HttpStateSurface::CodexCli);
     let issuer = format!("http://{}:{}", issuer_addr.ip(), issuer_addr.port());
 
     let tmp = tempdir()?;
@@ -394,7 +408,8 @@ async fn oauth_access_denied_missing_entitlement_blocks_login_with_clear_error()
 async fn oauth_access_denied_unknown_reason_uses_generic_error_page() -> Result<()> {
     skip_if_no_network!(Ok(()));
 
-    let (issuer_addr, _issuer_handle) = start_mock_issuer(WORKSPACE_ID_ALLOWED);
+    let (issuer_addr, _issuer_handle) =
+        start_mock_issuer(WORKSPACE_ID_ALLOWED, HttpStateSurface::CodexCli);
     let issuer = format!("http://{}:{}", issuer_addr.ip(), issuer_addr.port());
 
     let tmp = tempdir()?;
@@ -503,7 +518,8 @@ async fn falls_back_to_registered_fallback_port_when_default_port_is_in_use() ->
         })
     };
 
-    let (issuer_addr, _issuer_handle) = start_mock_issuer(WORKSPACE_ID_ALLOWED);
+    let (issuer_addr, _issuer_handle) =
+        start_mock_issuer(WORKSPACE_ID_ALLOWED, HttpStateSurface::CodexCli);
     let issuer = format!("http://{}:{}", issuer_addr.ip(), issuer_addr.port());
     let tmp = tempdir()?;
 
@@ -541,7 +557,8 @@ async fn falls_back_to_registered_fallback_port_when_default_port_is_in_use() ->
 async fn cancels_previous_login_server_when_port_is_in_use() -> Result<()> {
     skip_if_no_network!(Ok(()));
 
-    let (issuer_addr, _issuer_handle) = start_mock_issuer(WORKSPACE_ID_ALLOWED);
+    let (issuer_addr, _issuer_handle) =
+        start_mock_issuer(WORKSPACE_ID_ALLOWED, HttpStateSurface::CodexCli);
     let issuer = format!("http://{}:{}", issuer_addr.ip(), issuer_addr.port());
 
     let first_tmp = tempdir()?;
